Small config chenges

+++ rebuild.bat
@@ -1,3 +1,4 @@
 @echo off
-rd /Q /S public
+IF EXIST public rd /Q /S public
+IF EXIST resources\_gen rd /Q /S resources\_gen
 ..\hugo
